[support] support GetcuReone.Cdi.3.0.0 updates Fixes #39
