+++ GetcuReone.Cdi.WpfExtension/.build-tools/GetcuReone.Lib.props
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Project>
+
+    <Import Project="$(MSBuildThisFileDirectory)\GetcuReone.Common.props"/>
+    <Import Project="$(MSBuildThisFileDirectory)\getcureone-documentation.props"/>
+
+</Project>
