+++ GetcuReone.Cdi.WpfExtension/.build-tools/getcureone-documentation.props
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Project>
+
+    <!-- Live documentation settings-->
+    <PropertyGroup>
+        <DocumentationFile>$(MSBuildProjectName).xml</DocumentationFile>
+        <DocumentationMarkdown>$(MSBuildProjectDirectory)\API.md</DocumentationMarkdown>
+        <VsxmdAutoDeleteXml>True</VsxmdAutoDeleteXml>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Vsxmd" Version="1.4.5">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
+        </PackageReference>
+    </ItemGroup>
+
+</Project>
