+++ GetcuReone.Cdi.WpfExtension/Infrastructure/InfrastructureTests/InfrastructureTests.csproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <Import Project="..\..\.buildTools\GetcuReone.Common.props" />
+    <Import Project="..\..\.build-tools\GetcuReone.Common.props" />
 
     <PropertyGroup>
         <TargetFramework>netcoreapp3.1</TargetFramework>
